Separate frequency holder from whitelist
