--- generator/src/main/java/com/scottlogic/deg/generator/fieldspecs/Whitelist.java
@@ -1,8 +1,0 @@
-package com.scottlogic.deg.generator.fieldspecs;
-
-import java.util.Set;
-
-public interface Whitelist<T> {
-
-    Set<T> set();
-}
